backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.6
pandas==2.0.3
numpy==1.26.4
requests==2.32.3
pydantic==2.8.2
python-multipart==0.0.9
pytest==8.3.2
optuna==3.6.1
PyJWT==2.9.0
arq==0.26.3
redis==5.0.8
